requirements.txt
requests==2.33.1
rapidfuzz==3.14.3
playwright==1.58.0
playwright-stealth==2.0.2
firecrawl-py==4.22.1
pytest==8.4.2
